publications/requirements.txt
pyyaml>=6.0
requests>=2.31
playwright>=1.47